add event price

+++ src/main/java/GiftsBackend/Model/Product.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.List;
@@ -20,7 +21,7 @@ public class Product {
     @GeneratedValue
     private Long Id;
     private String Name;
-    private double Price;
+    private BigDecimal Price;
     private String ImageUrl;
     private LocalDate createdDate;
     @Convert(converter = Converter.class)
